+++ runner/src/server/plugins/engine/feedback/helpers.ts
@@ -1,0 +1,17 @@
+import atob from "atob";
+
+import { FeedbackContextInfo } from "./feedbackContextInfo";
+
+export function decodeFeedbackContextInfo(
+  encoded: string | Buffer | undefined
+): FeedbackContextInfo | void {
+  if (encoded) {
+    const decoded = JSON.parse(atob(encoded));
+
+    return new FeedbackContextInfo(
+      decoded.formTitle,
+      decoded.pageTitle,
+      decoded.url
+    );
+  }
+}
